+++ src/ledger/script.rs
@@ -102,11 +102,11 @@ impl Ledger {
             None => return Ok(()),
         };
 
-        let current_block_height = self.get_block_height() as u32;
-        let current_block_time = self.get_block_time(current_block_height as u64)? as u32;
+        let current_block_height = self.get_block_height();
+        let current_block_time = self.get_block_time(current_block_height)?;
 
         let tx_block_height = self.get_transaction_block_height(utxo.txid)?;
-        let tx_block_time = self.get_tx_block_height(utxo.txid) as u32;
+        let tx_block_time = self.get_tx_block_height(utxo.txid);
 
         let blocks_after = current_block_height - tx_block_height;
         let time_after = current_block_time - tx_block_time;
